fix: share search logic with pool details/creation pages
